Add files via upload
Rough version of the program built with Tensorflow-based FaceNet and RetinaFace models. I was yet to refactor this code before I moved over to PyTorch, and only accepts a target image and a video file to be processed. This version may still have some errors in it, but the repositories for these models are much better maintained than the PyTorch repositories I was able to find online, and likely yield better results. This program only has CPU support however, as I was unable to learn how to use Windows Subsystem for Linux in time.